LayoutTests/http/tests/local/formdata/send-form-data-with-sliced-file-expected.txt

CONSOLE MESSAGE: line 76: TypeError: 'undefined' is not a function (evaluating 'file.webkitSlice(dataList[i]['start'], dataList[i]['start'] + dataList[i]['length'])')

Test for sending FormData with sliced files via XMLHttpRequest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Sending FormData containing one sliced file with empty name:

Sending FormData containing one sliced file:
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
